refactor: api directory
