--- src/socket/room.ts
@@ -1,11 +1,0 @@
-"use client";
-
-import { io } from "socket.io-client";
-
-const dev = process.env.NODE_ENV !== "production";
-
-const host = dev ? "localhost:3000" : process.env.NEXT_PUBLIC_HOSTNAME;
-
-export const socket = io(`http://${host}`, {
-  path: "/socket.io/",
-});
